+++ packages/styles/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 10.29.10 (2024-02-13)
+
+**Note:** Version bump only for package @wfp/styles
+
+
+
+
+
+## 10.29.9 (2024-02-13)
+
+**Note:** Version bump only for package @wfp/styles
+
+
+
+
+
+## 10.29.8 (2024-01-26)
+
+**Note:** Version bump only for package @wfp/styles
+
+
+
+
+
+## 10.29.7 (2024-01-23)
+
+**Note:** Version bump only for package @wfp/styles
+
+
+
+
+
 ## 10.29.8 (2024-01-22)
 
 **Note:** Version bump only for package @wfp/styles